+++ src/Silex/ControllerCollection.php
@@ -12,7 +12,6 @@
 namespace Silex;
 
 use Symfony\Component\Routing\RouteCollection;
-use Symfony\Component\Routing\Route;
 use Silex\Controller;
 
 /**
@@ -28,6 +27,17 @@ use Silex\Controller;
 class ControllerCollection
 {
     protected $controllers = array();
+    protected $defaultRoute;
+
+    /**
+     * Constructor.
+     *
+     * @param Route $route
+     */
+    public function __construct()
+    {
+        $this->defaultRoute = new Route('');
+    }
 
     /**
      * Maps a pattern to a callable.
@@ -41,9 +51,11 @@ class ControllerCollection
      */
     public function match($pattern, $to)
     {
-        $route = new Route($pattern, array('_controller' => $to));
-        $controller = new Controller($route);
-        $this->add($controller);
+        $route = clone $this->defaultRoute;
+        $route->setPattern($pattern);
+        $route->setDefault('_controller', $to);
+
+        $this->controllers[] = $controller = new Controller($route);
 
         return $controller;
     }
@@ -101,13 +113,129 @@ class ControllerCollection
     }
 
     /**
-     * Adds a controller to the staging area.
+     * Sets the requirement for a route variable.
      *
-     * @param Controller $controller
+     * @param string $variable The variable name
+     * @param string $regexp   The regexp to apply
+     *
+     * @return Controller $this The current Controller instance
      */
-    public function add(Controller $controller)
+    public function assert($variable, $regexp)
     {
-        $this->controllers[] = $controller;
+        $this->defaultRoute->assert($variable, $regexp);
+
+        foreach ($this->controllers as $controller) {
+            $controller->assert($variable, $regexp);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets the default value for a route variable.
+     *
+     * @param string $variable The variable name
+     * @param mixed  $default  The default value
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function value($variable, $default)
+    {
+        $this->defaultRoute->value($variable, $default);
+
+        foreach ($this->controllers as $controller) {
+            $controller->value($variable, $default);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets a converter for a route variable.
+     *
+     * @param string $variable The variable name
+     * @param mixed  $callback A PHP callback that converts the original value
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function convert($variable, $callback)
+    {
+        $this->defaultRoute->convert($variable, $callback);
+
+        foreach ($this->controllers as $controller) {
+            $controller->convert($variable, $callback);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets the requirement for the HTTP method.
+     *
+     * @param string $method The HTTP method name. Multiple methods can be supplied, delimited by a pipe character '|', eg. 'GET|POST'
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function method($method)
+    {
+        $this->defaultRoute->method($method);
+
+        foreach ($this->controllers as $controller) {
+            $controller->method($method);
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets the requirement of HTTP (no HTTPS) on this controller.
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function requireHttp()
+    {
+        $this->defaultRoute->requireHttp();
+
+        foreach ($this->controllers as $controller) {
+            $controller->requireHttp();
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets the requirement of HTTPS on this controller.
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function requireHttps()
+    {
+        $this->defaultRoute->requireHttps();
+
+        foreach ($this->controllers as $controller) {
+            $controller->requireHttps();
+        }
+
+        return $this;
+    }
+
+    /**
+     * Sets a callback to handle before triggering the route callback.
+     * (a.k.a. "Route Middleware")
+     *
+     * @param mixed $callback A PHP callback to be triggered when the Route is matched, just before the route callback
+     *
+     * @return Controller $this The current Controller instance
+     */
+    public function middleware($callback)
+    {
+        $this->defaultRoute->middleware($callback);
+
+        foreach ($this->controllers as $controller) {
+            $controller->middleware($callback);
+        }
+
+        return $this;
     }
 
     /**