+++ doc/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 This changelog references all backward incompatibilities as we introduce them:
 
+* **2012-05-20**: Added a way to define settings on a controller collection.
+
 * **2012-05-20**: The Request instance is not available anymore from the
   Application after it has been handled.